part-2/code/requirements.txt
azure-identity
azure-ai-projects
jsonref
python-dotenv
pandas
tabulate